refactor: split application routes
